% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winratio_function.R
\name{WinRatio_sampsize}
\alias{WinRatio_sampsize}
\title{Sample Size Estimation for the Win Ratio}
\usage{
WinRatio_sampsize(
  n_arm_1,
  n_arm_2,
  alpha = 0.05,
  WinRatio1 = NULL,
  p1 = NULL,
  WinRatio2 = NULL,
  p2 = NULL,
  corr = 0,
  cens_prop = 0,
  ties = 0,
  ties_dist = NULL,
  n.iter = 500,
  seed = 1000,
  boot = 500,
  print.iter = FALSE
)
}
\arguments{
\item{n_arm_1}{The sample size in arm 1}

\item{n_arm_2}{The sample size in arm 2}

\item{alpha}{The two-sided type I error rate}

\item{WinRatio1}{The specified Win Ratio for ratio of Arm 1 wins to Arm 1 losses on the first endpoint, either p1 or WinRatio1 should be specified}

\item{p1}{The probability of success in arm 1 for the first endpoint, either p1 or WinRatio1 should be specified}

\item{WinRatio2}{The specified Win Ratio for ratio of Arm 1 wins to Arm 1 losses on the second endpoint, either p2 or WinRatio2 can be specified. Do not include if only assuming one endpoint}

\item{p2}{The probability of success in arm 1 for the second endpoint, either p2 or WinRatio2 can be specified. Do not include if only assuming one endpoint}

\item{corr}{The correlation of endpoints. Must be specified if p2 or WinRatio2 is specified.}

\item{cens_prop}{The proportion of administrative censoring on the first endpoint}

\item{ties}{The proportion of ties on the second endpoint}

\item{ties_dist}{Equal to either "best" if the better ranks are tied or "worst" if worse ranks are tied}

\item{n.iter}{The number of simulations}

\item{seed}{The seed for random number generation}

\item{boot}{The number of bootstrap samples within each simulation}

\item{print.iter}{If TRUE, each simulation number will be printed after completion}
}
\value{
A list with components including the median, 25th percentile, 75th percentile, mean, and standard deviation of the true win ratios assuming complete data and the
observed win ratios assuming censoring and/or ties across simulations;
and the power of the proposed sample size under specified assumptions
}
\description{
A simulation-based sample size calculation approach, relying on the relationship between the win ratio and the rank distribution
}
\examples{
sim_example<-WinRatio_sampsize(n_arm_1=10,n_arm_2=10,alpha=0.05,p1=0.6,n.iter=100,seed=1234,boot=100)
}
